# trace network: use same code pattern as other networking gadgets