--- a/integration/ig/non-k8s/trace_network_test.go
+++ b/integration/ig/non-k8s/trace_network_test.go
@@ -40,6 +40,9 @@ func TestTraceNetwork(t *testing.T) {
 							Container: cn,
 						},
 					},
+					Comm:       "curl",
+					Uid:        0,
+					Gid:        0,
 					PktType:    "OUTGOING",
 					Proto:      "tcp",
 					Port:       80,
@@ -52,6 +55,9 @@ func TestTraceNetwork(t *testing.T) {
 							Container: cn,
 						},
 					},
+					Comm:       "nginx",
+					Uid:        0, // different nginx cmdline seems to cause different uid
+					Gid:        0,
 					PktType:    "HOST",
 					Proto:      "tcp",
 					Port:       80,
@@ -61,7 +67,10 @@ func TestTraceNetwork(t *testing.T) {
 
 			normalize := func(e *networkTypes.Event) {
 				e.Timestamp = 0
+				e.MountNsID = 0
 				e.NetNsID = 0
+				e.Pid = 0
+				e.Tid = 0
 			}
 
 			return ExpectEntriesToMatch(output, normalize, expectedEntries...)