--- a/pkg/gadgets/trace/network/tracer/bpf/graphmap.c
+++ b/pkg/gadgets/trace/network/tracer/bpf/graphmap.c
@@ -1,4 +1,0 @@
-// SPDX-License-Identifier: GPL-2.0
-/* Copyright (c) 2022 The Inspektor Gadget authors */
-
-#include "graphmap.h"