… pkg/gadget-collection/gadgets/trace/network/enricher.go
… pkg/gadgets/trace/network/tracer/bpf/graph.h
… pkg/gadgets/trace/network/tracer/bpf/graphmap.h
… pkg/gadgets/trace/network/tracer/graph_bpfel.go
… pkg/gadgets/trace/network/tracer/graph_bpfel.o
… pkg/gadgets/trace/network/tracer/graphmap_bpfel.go
… pkg/gadgets/trace/network/tracer/graphmap_bpfel.o
… pkg/gadgets/trace/network/tracer/network_bpfel.go
… pkg/gadgets/trace/network/tracer/network_bpfel.o