# trace network: use same code pattern as other networking gadgets

- Use pkg/gadgets/internal/networktracer package instead of 
  - Use a perf ring buffer like other tracing gadgets instead of a custom hash map polled every second. This can be slower with a lot of new connections but I am hoping we can get the performances back with the Prometheus work.
- Use kubeipresolver instead of its own enricher
- Use socket enricher

First part of https://github.com/inspektor-gadget/inspektor-gadget/issues/1612
Refactoring suggested in https://github.com/inspektor-gadget/inspektor-gadget/pull/1430#issuecomment-1470794513
Fixes https://github.com/inspektor-gadget/inspektor-gadget/issues/1334
Supersedes https://github.com/inspektor-gadget/inspektor-gadget/pull/1336

Depends on:
- [x] https://github.com/inspektor-gadget/inspektor-gadget/pull/1636
- [x] https://github.com/inspektor-gadget/inspektor-gadget/pull/1680

I will rebase this PR once the dependencies are merged.

## How to use

[ describe what reviewers need to do in order to validate this PR ]

## Testing done

[Describe the testing you have done before submitting this PR. Please include both the commands you issued as well as the output you got.]