@alban 2023-05-30
I am testing this in two ways:

1. With trace network:
```
$ ./kubectl-gadget trace network -o json|grep tcp
{"node":"minikube-containerd","namespace":"default","pod":"shell02","container":"shell02","timestamp":1685466705588526838,"type":"normal","mountnsid":4026533430,"netnsid":4026533256,"pid":255628,"tid":255628,"comm":"nc","pktType":"OUTGOING","proto":"tcp","port":9091,"podHostIP":"192.168.49.2","podIP":"10.244.0.7","podLabels":{"run":"shell02"},"remoteKind":"svc","remoteAddr":"10.99.179.186","remoteName":"shell","remoteNamespace":"default","remoteLabels":{"run":"shell"}}
```

2. With advise network-policy
```
$ ./kubectl-gadget advise network-policy monitor|grep tcp
{"node":"minikube-containerd","namespace":"default","pod":"shell02","timestamp":1685466740985560531,"type":"normal","mountnsid":4026533430,"netnsid":4026533256,"pid":256061,"tid":256061,"comm":"nc","pktType":"OUTGOING","proto":"tcp","port":9091,"remoteAddr":"10.99.179.186"}
```

As you can see, KubeIPResolver enriches the event correctly with 'trace network' but not with 'advise network-policy'. This is because:
1. 'trace network' uses the operators to enrich the events but 'advise network-policy' still uses pkg/gadget-collection/gadgets/trace/network/gadget.go.
2. In this PR, I no longer pass an enricher interface to netTracer.NewTracer(). And I removed the file pkg/gadget-collection/gadgets/trace/network/enricher.go since it is a duplicate of pkg/operators/kubeipresolver/kubeipresolver.go.

Is there a simple way I could get a reference to the KubeIPResolver from pkg/gadget-collection/gadgets/trace/network/gadget.go in order to enrich the events from there?

Alternatively, cmd/kubectl-gadget/advise/network-policy.go could be rewritten to use the gRPC interface instead of the CR. But I would like to avoid refactoring too much in a single PR.

cc @flyth