@alban 2023-05-31
> Is there a simple way I could get a reference to the KubeIPResolver from pkg/gadget-collection/gadgets/trace/network/gadget.go in order to enrich the events from there?
> 
> Alternatively, cmd/kubectl-gadget/advise/network-policy.go could be rewritten to use the gRPC interface instead of the CR. But I would like to avoid refactoring too much in a single PR.

Updated by getting the KubeIPResolver operator with GetRaw(), based on an idea from @flyth.