@alban 2023-06-05
Rebased and added the uid/gid feature like it was done for other gadgets.