@alban 2023-06-06
> Note that this PR worsens the performances. It is partially expected because it sends all events to userspace instead of aggregating results in a hash map. I was planning to reuse the work for Prometheus in ebpf (https://github.com/inspektor-gadget/inspektor-gadget/issues/1598) to make it fast again but reusing the general code pattern instead of having something specific for trace-network.
> 
> However, I am not sure why the benchmarks are so bad because so far the benchmarks don't generate any events but just load+unload the ebpf program

```
$ go test -exec sudo -ldflags="-s=false" -bench='^BenchmarkAllGadgetsWithContainers$/^container1$/trace-network' -run=Benchmark ./internal/benchmarks/... -count 10 | tee patched.bench
$ git checkout main
$ go mod vendor
$ go test -exec sudo -ldflags="-s=false" -bench='^BenchmarkAllGadgetsWithContainers$/^container1$/trace-network' -run=Benchmark ./internal/benchmarks/... -count 10 | tee main.bench
$ benchstat main.bench patched.bench|tee benchstat.log
goos: linux
goarch: amd64
pkg: github.com/inspektor-gadget/inspektor-gadget/internal/benchmarks
cpu: Intel(R) Core(TM) i7-6500U CPU @ 2.50GHz
                                                    │  main.bench  │              patched.bench               │
                                                    │    sec/op    │     sec/op      vs base                  │
AllGadgetsWithContainers/container1/trace-network-4   12.43m ± 11%   1220.34m ± 47%  +9715.57% (p=0.000 n=10)
```

This PR decreases the performances of trace network by +9715.57% and bring it to the same "bad" level as trace dns.

This is due to the socket enricher, which was not used in the trace network gadget before the PR:

```
$ sudo /usr/share/bcc/tools/profile -p $(pidof benchmarks.test)
...
    github.com/cilium/ebpf/btf.LoadKernelSpec
    github.com/cilium/ebpf.findTargetInKernel
    github.com/cilium/ebpf.newProgramWithOptions
    github.com/cilium/ebpf.(*collectionLoader).loadProgram
    github.com/cilium/ebpf.(*CollectionSpec).LoadAndAssign.func1
    github.com/cilium/ebpf.assignValues
    github.com/cilium/ebpf.(*CollectionSpec).LoadAndAssign
    github.com/inspektor-gadget/inspektor-gadget/pkg/gadgets/internal/socketenricher.(*SocketEnricher).start
    github.com/inspektor-gadget/inspektor-gadget/pkg/gadgets/internal/socketenricher.NewSocketEnricher
    github.com/inspektor-gadget/inspektor-gadget/pkg/gadgets/internal/networktracer.NewTracer[...]
    github.com/inspektor-gadget/inspektor-gadget/pkg/gadgets/trace/network/tracer.(*Tracer).install
    github.com/inspektor-gadget/inspektor-gadget/pkg/gadgets/trace/network/tracer.(*Tracer).Init
    github.com/inspektor-gadget/inspektor-gadget/pkg/runtime/local.(*Runtime).RunGadget
    github.com/inspektor-gadget/inspektor-gadget/internal/benchmarks.BenchmarkAllGadgetsWithContainers.func1.1
    testing.(*B).runN
    testing.(*B).run1.func1
    runtime.goexit.abi0
...
```

Since the time is mostly spent in the socket enricher's LoadAndAssign and LoadKernelSpec, we can do the following optimization for gadgettracermanager:
* Load the socket enricher when the process starts, but without attaching the kprobes
* When the socket enricher is used by a gadget, initialize it (~RewriteConstants) and attach the kprobes
* When the socket enricher is not used anymore, detach the kprobes but don't release the eBPF programs so it can be reused for the next gadget.

This would reduce the time to load a gadget using the socket enricher by about 1.2 seconds, so I think it is worth it for a better user experience. This is unrelated to this PR though, since it is the same problem in trace dns, sni, tcpdrop, etc.